web/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.1
idna==3.4
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.5.0
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
